order-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
requests